+++ src/main/java/FriendlyTool/command/CommandTypes.java
@@ -1,3 +1,5 @@
+package FriendlyTool.command;
+
 public enum CommandTypes {
     TODO, DEADLINE, EVENT, LIST, BYE, MARK, UNMARK, DELETE
 }
