+++ src/main/java/FriendlyTool/main/Date.java
@@ -1,7 +1,7 @@
-import java.time.LocalDate;
+package FriendlyTool.main;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 public class Date {
     LocalDateTime date;
