+++ src/main/java/FriendlyTool/main/ftException.java
@@ -1,3 +1,5 @@
+package FriendlyTool.main;
+
 public class ftException extends  Exception {
     public ftException(String message) {
         super(message);
