+++ src/main/java/FriendlyTool/task/Task.java
@@ -1,0 +1,36 @@
+package FriendlyTool.task;
+
+public class Task {
+    private String name;
+    private boolean isDone;
+
+    public Task(String name, boolean isDone) {
+        this.name = name;
+        this.isDone = isDone;
+    }
+
+    public void mark() {
+        isDone = true;
+    }
+
+    public void unmark() {
+        isDone = false;
+    }
+
+    public boolean isDone() {
+        return this.isDone;
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + (isDone ? "X" : " ") + "] " + name;
+    }
+    public String toSaveFormat() {
+        return "| " + (isDone ? "true" : "false") + " | " + name;
+    }
+}
+
